@@ -6,88 +6,97 @@
   "scripts": {
     "clean": "ionic-app-scripts clean",
     "build": "ionic-app-scripts build",
+    "lint": "ionic-app-scripts lint",
     "ionic:build": "ionic-app-scripts build",
-    "ionic:serve": "ionic-app-scripts serve"
+    "ionic:serve": "ionic-app-scripts serve --watch ./config/watch.config.js"
   },
-  "dependencies": {
-    "@angular/common": "4.0.2",
-    "@angular/compiler": "4.0.2",
-    "@angular/compiler-cli": "4.0.2",
-    "@angular/core": "4.0.2",
-    "@angular/forms": "4.0.2",
-    "@angular/http": "4.0.2",
-    "@angular/platform-browser": "4.0.2",
-    "@angular/platform-browser-dynamic": "4.0.2",
-    "@ionic-native/core": "3.6.0",
-    "@ionic-native/splash-screen": "3.6.0",
-    "@ionic-native/status-bar": "3.6.0",
-    "@ionic/storage": "2.0.1",
-    "ionic-angular": "3.1.0",
-    "ionicons": "3.0.0",
-    "rxjs": "5.1.1",
-    "sw-toolbox": "3.4.0",
-    "zone.js": "^0.8.5",
-
-    "@angular/platform-server": "4.0.2",
-    "@angular/animations": "4.0.2",
-    "@ionic-native/app-version": "^3.6.0",
-    "@ionic-native/barcode-scanner": "^3.6.0",
-    "@ionic-native/camera": "^3.6.0",
-    "@ionic-native/geolocation": "^3.6.0",
-    "@ionic-native/network": "^3.6.0",
-    "@ionic-native/onesignal": "^3.6.0",
-    "@ionic-native/photo-viewer": "^3.6.0",
-    "angular2-qrcode": "git+https://github.com/dapriett/angular2-qrcode",
-    "elementtree": "^0.1.7",
-    "flag-icon-css": "^2.8.0",
-    "font-awesome": "^4.7.0",
-    "gl-ionic2-env-configuration": "0.0.18",
-    "imgcache.js": "^1.1.1",
-    "ionic-cache": "^1.1.2",
-    "ionic2-rating": "^1.2.0",
-    "plist": "^2.0.1",
-    "sprintf-js": "^1.0.3",
-    "ts-md5": "^1.2.0"
-  },
-  "devDependencies": {
-    "@ionic/app-scripts": "1.3.4",
-    "typescript": "~2.2.1"
-  },
-  "cordovaPlugins": [
-    "ionic-plugin-keyboard",
-    "cordova-plugin-whitelist",
-    "cordova-plugin-statusbar",
-    "cordova-plugin-device",
-    "cordova-plugin-splashscreen",
-    "cordova-plugin-file",
-    "cordova-plugin-compat",
-    "cordova-plugin-file-transfer",
-    "cordova-plugin-console",
-    "phonegap-plugin-barcodescanner",
-    "cordova-plugin-app-version",
-    "onesignal-cordova-plugin",
-    "cordova-plugin-network-information",
-    "com-sarriaroman-photoviewer",
-    "cordova-plugin-settings-hook"
-  ],
-  "description": "partymeister-companion-app: An Ionic project",
   "config": {
-    "ionic_bundler": "webpack",
-    "ionic_source_map": "source-map",
+    "ionic_watch": "./config/watch.config.js",
     "ionic_copy": "./config/copy.config.js",
     "ionic_sass": "./config/sass.config.js",
     "ionic_webpack": "./config/webpack.config.js"
   },
-  "cordovaPlatforms": [
-    {
-      "platform": "android",
-      "version": "",
-      "locator": "android"
+  "dependencies": {
+    "@angular/animations": "5.0.0",
+    "@angular/common": "5.0.0",
+    "@angular/compiler": "5.0.0",
+    "@angular/compiler-cli": "5.0.0",
+    "@angular/core": "5.0.0",
+    "@angular/forms": "5.0.0",
+    "@angular/http": "5.0.0",
+    "@angular/platform-browser": "5.0.0",
+    "@angular/platform-browser-dynamic": "5.0.0",
+    "@fortawesome/angular-fontawesome": "^0.1.0-3",
+    "@ionic-native/app-version": "^4.5.3",
+    "@ionic-native/barcode-scanner": "^4.5.3",
+    "@ionic-native/camera": "^4.5.3",
+    "@ionic-native/core": "^4.5.1",
+    "@ionic-native/file": "^4.5.3",
+    "@ionic-native/geolocation": "^4.5.3",
+    "@ionic-native/network": "^4.5.3",
+    "@ionic-native/onesignal": "^4.5.3",
+    "@ionic-native/photo-viewer": "^4.5.3",
+    "@ionic-native/splash-screen": "^4.5.3",
+    "@ionic-native/status-bar": "^4.5.3",
+    "@ionic/storage": "2.1.3",
+    "angular2-qrcode": "^2.0.1",
+    "com-sarriaroman-photoviewer": "^1.1.15",
+    "cordova-android": "^7.0.0",
+    "cordova-ios": "4.5.4",
+    "cordova-plugin-app-version": "^0.1.9",
+    "cordova-plugin-device": "^1.1.7",
+    "cordova-plugin-file": "^5.0.0",
+    "cordova-plugin-file-transfer": "^1.7.1",
+    "cordova-plugin-inappbrowser": "^2.0.2",
+    "cordova-plugin-network-information": "^1.3.4",
+    "cordova-plugin-settings-hook": "^0.2.7",
+    "cordova-plugin-splashscreen": "^4.1.0",
+    "cordova-plugin-statusbar": "^2.4.1",
+    "cordova-plugin-whitelist": "^1.3.3",
+    "elementtree": "^0.1.7",
+    "flag-icon-css": "^2.9.0",
+    "gl-ionic2-env-configuration": "0.0.18",
+    "ionic-angular": "3.9.2",
+    "ionic-cache": "^2.0.5",
+    "ionic-image-loader": "^5.0.0",
+    "ionic-plugin-keyboard": "^2.2.1",
+    "ionic2-rating": "^1.2.2",
+    "ionicons": "3.0.0",
+    "onesignal-cordova-plugin": "^2.2.5",
+    "phonegap-plugin-barcodescanner": "^6.0.8",
+    "plist": "^2.0.1",
+    "rxjs": "5.5.2",
+    "sw-toolbox": "3.6.0",
+    "ts-md5": "^1.2.4",
+    "zone.js": "0.8.18"
+  },
+  "devDependencies": {
+    "@ionic/app-scripts": "^3.1.8",
+    "typescript": "2.4.2"
+  },
+  "cordova": {
+    "plugins": {
+      "cordova-plugin-inappbrowser": {},
+      "cordova-plugin-network-information": {},
+      "cordova-plugin-statusbar": {},
+      "ionic-plugin-keyboard": {},
+      "onesignal-cordova-plugin": {},
+      "cordova-plugin-file": {},
+      "cordova-plugin-file-transfer": {},
+      "cordova-plugin-device": {},
+      "cordova-plugin-whitelist": {},
+      "cordova-plugin-splashscreen": {},
+      "cordova-plugin-app-version": {},
+      "cordova-plugin-settings-hook": {},
+      "phonegap-plugin-barcodescanner": {
+        "CAMERA_USAGE_DESCRIPTION": " "
+      },
+      "com-sarriaroman-photoviewer": {}
     },
-    {
-      "platform": "ios",
-      "version": "",
-      "locator": "ios"
-    }
-  ]
+    "platforms": [
+      "android",
+      "ios"
+    ]
+  },
+  "description": "partymeister-companion-app: An Ionic project"
 }
