@@ -15,13 +15,14 @@ module.exports = {
         dest: '{{WWW}}/assets/fonts'
     },
     copyPolyfills: {
-        src: ['{{ROOT}}/node_modules/ionic-angular/polyfills/polyfills.js'],
+        src: [`{{ROOT}}/node_modules/ionic-angular/polyfills/${process.env.IONIC_POLYFILL_FILE_NAME}`],
         dest: '{{BUILD}}'
     },
     copySwToolbox: {
         src: ['{{ROOT}}/node_modules/sw-toolbox/sw-toolbox.js'],
         dest: '{{BUILD}}'
     },
+
     copyFontAwesome: {
         src: '{{ROOT}}/node_modules/font-awesome/fonts/*',
         dest: '{{WWW}}/fonts/'
