@@ -1,33 +1,35 @@
-// Set the `ENV` global variable to be used in the app.
+var chalk = require("chalk");
+var fs = require('fs');
 var path = require('path');
-var webpack = require('webpack');
+var useDefaultConfig = require('@ionic/app-scripts/config/webpack.config.js');
 
-var projectRootDir = process.env.IONIC_ROOT_DIR;
-var appScriptsDir = process.env.IONIC_APP_SCRIPTS_DIR;
+var env = process.env.IONIC_ENV;
 
-var config = require(path.join(appScriptsDir, 'config', 'webpack.config.js'));
+useDefaultConfig.prod.resolve.alias = {
+    "@app/env": path.resolve(environmentPath('prod'))
+};
 
-var env = process.env.IONIC_ENV || 'dev';
-var envVars;
+useDefaultConfig.dev.resolve.alias = {
+    "@app/env": path.resolve(environmentPath('dev'))
+};
 
-try {
-    envVars = require(path.join(projectRootDir, 'src/app/environment-variables', env + '.json'));
-} catch(e) {
-    envVars = {};
+if (env !== 'prod' && env !== 'dev') {
+    // Default to dev config
+    useDefaultConfig[env] = useDefaultConfig.dev;
+    useDefaultConfig[env].resolve.alias = {
+        "@app/env": path.resolve(environmentPath(env))
+    };
 }
 
-config.plugins = config.plugins || [];
-config.plugins.push(
-    new webpack.DefinePlugin({
-        ENV: Object.assign(envVars, {
-            environment: JSON.stringify(env)
-        })
-    })
-);
-
-if(env === 'prod') {
-    // This helps ensure the builds are consistent if source hasn't changed:
-    config.plugins.push(new webpack.optimize.OccurrenceOrderPlugin());
+function environmentPath(env) {
+    var filePath = './src/environments/environment' + (env === 'prod' ? '' : '.' + env) + '.ts';
+    if (!fs.existsSync(filePath)) {
+        console.log(chalk.red('\n' + filePath + ' does not exist!'));
+    } else {
+        return filePath;
+    }
 }
 
-module.exports = config;
+module.exports = function () {
+    return useDefaultConfig;
+};
