@@ -1,9 +1,13 @@
 import {Injectable} from '@angular/core';
-import {Http, Response, Headers, RequestOptions} from '@angular/http';
 import {SettingsProvider} from './settings';
 import 'rxjs/add/operator/map';
 import {StorageProvider} from '../providers/storage';
 import {User} from '../models/user';
+import {HttpClient, HttpHeaders} from "@angular/common/http";
+
+const httpOptions = {
+    headers: new HttpHeaders({'Content-Type': 'application/json', 'token': SettingsProvider.variables.API_TOKEN})
+};
 
 @Injectable()
 export class AuthProvider {
@@ -11,7 +15,7 @@ export class AuthProvider {
     private authenticated: boolean = false;
     public user: User;
 
-    constructor(public http: Http, private storageProvider: StorageProvider) {
+    constructor(public http: HttpClient, private storageProvider: StorageProvider) {
         this.storageProvider.get('user').then(res => {
             if (res != null) {
                 this.user = <User><any>res;
@@ -46,10 +50,8 @@ export class AuthProvider {
 
     loginRequest(data) {
         let bodyString = JSON.stringify(data); // Stringify payload
-        let headers = new Headers({'Content-Type': 'application/json', 'token': SettingsProvider.variables.API_TOKEN}); // ... Set content type to JSON
-        let options = new RequestOptions({headers: headers}); // Create a request option
 
-        return this.http.post(SettingsProvider.variables.LOGIN_API, bodyString, options)
+        return this.http.post(SettingsProvider.variables.LOGIN_API, bodyString, httpOptions)
             .map(res => {
                 return res;
             });
@@ -57,10 +59,10 @@ export class AuthProvider {
 
     registrationRequest(data) {
         let bodyString = JSON.stringify(data); // Stringify payload
-        let headers = new Headers({'Content-Type': 'application/json', 'token': SettingsProvider.variables.API_TOKEN}); // ... Set content type to JSON
-        let options = new RequestOptions({headers: headers}); // Create a request option
 
-        return this.http.post(SettingsProvider.variables.REGISTRATION_API, bodyString, options).map(res => { return res;} );
+        return this.http.post(SettingsProvider.variables.REGISTRATION_API, bodyString, httpOptions).map(res => {
+            return res;
+        });
     }
 
 }
