@@ -1,12 +1,17 @@
 import {Injectable} from '@angular/core';
 import 'rxjs/add/operator/map';
 import {SettingsProvider} from './settings';
-import {Http, Response, Headers, RequestOptions} from '@angular/http';
+import {HttpClient, HttpHeaders} from "@angular/common/http";
+
+const httpOptions = {
+    headers: new HttpHeaders({'Content-Type': 'application/json', 'token': SettingsProvider.variables.API_TOKEN})
+};
+
 
 @Injectable()
 export class VisitorProvider {
 
-    constructor(public http: Http) {
+    constructor(public http: HttpClient) {
     }
 
     filterItems(visitors, searchTerm) {
@@ -28,10 +33,10 @@ export class VisitorProvider {
 
     signupRequest(data) {
         let bodyString = JSON.stringify(data); // Stringify payload
-        let headers = new Headers({'Content-Type': 'application/json', 'token': SettingsProvider.variables.API_TOKEN}); // ... Set content type to JSON
-        let options = new RequestOptions({headers: headers}); // Create a request option
 
-        return this.http.post(SettingsProvider.variables.VISITOR_SIGNUP_API, bodyString, options).map(res => { return res} );
+        return this.http.post(SettingsProvider.variables.VISITOR_SIGNUP_API, bodyString, httpOptions).map(res => {
+            return res
+        });
     }
 
 }
